requirement.txt
fastapi
uvicorn
datetime
mysql-connector-python
